Adjust MainDisplay CSS
Changes the className of the main textbox to "MainDisplay" and adjusts the CSS.

The idea here is to create a MainDisplay box that can have items appended to it. This would be things like text, images, etc. This will become the home of the chat between the User and Harold.